Reduce Balance / BlockNumber Integer Width
This patch reduces the width of all of our storage variables to consistently use 96 bits for amounts internally, and 32 bits for blocks. This is consistent with how we store Checkpoints, but prevents us from consistently having to cast integer sizes internally. Instead, each public function that accepts or returns uint256's (due to the Erc-20 spec) simply does the casting directly on its inputs. This leads to consistently cleaner code.