Add Token Deployment Script (#34)
* Add Token Deployment Script

This patch adds saddle scripts to deploy, verify and match a cToken from a JSON configuration passed in on the command line. The goal is to make life very easy for deployment, specifically with builds from release versions of the Compound Protocol. Note: this does not _force_ developers to use release versions, but it means that developers can choose to deploy from versions of Compound that were audited. I've added significant usage notes to the README.

* Add Docker Builds

This patch adds Docker builds to our CI pipeline. This is the last step in letting the community run verified deployments. Every time a commit is pushed, an image and built and pushed to Docker Hub tagged with the commit (and if the commit is tagged, that will be included as a tag for Docker Hub as well).